feat: Add PAGESIZE enum for material page size
Add a new enum PAGESIZE with a constant MATERIAL_PAGE_SIZE set to 15. This enum will be used to define the number of materials to display per page in the brainstorming section. This change allows for easier customization of the page size and improves the user experience by providing a consistent number of materials per page.